refactor(backend): move db/auth.rs to more appropriate files (it doesn't touch the db)
