+++ backend/api/src/db.rs
@@ -1,4 +1,3 @@
-pub mod auth;
 pub mod user;
 
 use sqlx::postgres::{PgConnectOptions, PgPool, PgPoolOptions};
